9-01-2026/backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
faiss-cpu
sentence-transformers
pymongo
python-jose
passlib
google-generativeai
